--- HW rev.1 ---
in mj838.c/.h:
- removed solid-state relay control GPIO initialization in _GPIOInit()
- added (float) typecast to fix compiler warnings

in autocharge.c:
- added typecast for __AutoCharge.__SW_C* to avoid compiler warnings
- commented out solid-state relay control GPIO initialization in autocharge_ctor()

in autodrive.c
- added (float) typecast to fix compiler warnings

in mj838_zerocross.c:
- added (float) & other typecasts to fix compiler warnings

start of code conversion to MJ838 HW rev.1

- seperated MJ8x8 clone - rev.5 - TCAN334 - poc.brd(generic) BOMs from device specific

minor corrections on BOM

in GnZC - POC.brd/sch:
- changed 3V3 power symbol

mj838 brd & sch:
- first paste of GnZC sheets into base mj8x8 schematic, thus becoming mj838
- netclass corrections
- recreated blank brd
- fixed most ERCs
- edit & cleanup of sch
- renamed to hw rev 1
- added BAT+ connection via diode/0R resistor towards VDC
- added artificial load sheet
- added R3119x - Input Voltage Detector for 3V3 LDO control on smoothing cap charge
- added MB85RC16 - FeRAM Memory
- added testpoints to reach of the STM32 GPIOs
- added PAC1954 - Power Monitor
- added SHT40-AD1F - Humidity Sensor
- added TLC59208F - LED Driver and incorporated said driver into schematic
- created BOM & gerbers

mj838 housing dwg:
- drew cigarette-pack sized box

in GnZC - POC.brd/sch:
- updated conceptual overview to include 3V3 LDO
- moved LDO reverse protection diode into 3V3 sheet
- changed 3V3 power symbol

mj838 project:
- added GnZC files

mj838.dwg:
- added blank dwg

+++ mj8x8/Core/Inc/mj838/autodrive.c
@@ -89,7 +89,7 @@ static void _Do(void)  // this actually runs the AutoDrive application
 {
 	__AutoDrive._WheelFrequency = Device->ZeroCross->GetZCFrequency() / POLE_COUNT;  // ZeroCross signal frequency to wheel RPS
 	__AutoDrive.mps.Float = __AutoDrive._WheelFrequency * WHEEL_CIRCUMFERENCE;	// wheel frequency to m/s
-	__AutoDrive.kph.Float = __AutoDrive.mps.Float * 3.6;  // m/s to km/h
+	__AutoDrive.kph.Float = (float) (__AutoDrive.mps.Float * 3.6);  // m/s to km/h
 	__AutoDrive.m.Float += __AutoDrive.mps.Float * (float) ((__HAL_TIM_GET_AUTORELOAD(&htim3) + 1) / 10000.0);  // distance, mps * measurement interval
 
 #if SIGNAL_GENERATOR_INPUT
